+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - **[Fix]** Fix panic on invalid cap style.
 - **[Fix]** Fix panic on unmatched morph shape record pair.
 - **[Fix]** Fix panic on invalid GIF or PNG header.
+- **[Fix]** Fix panic on text definition with invalid index or advance bits.
 
 ## Typescript